refactor: small refactor on artwork scene
Moved helper functions to helpers.ts file and moved placeholder components to their own files to reduce the size of the Artwork.tsx Screen file
