+++ src/app/Scenes/Artwork/Components/BelowTheFoldPlaceholder.tsx
@@ -1,0 +1,26 @@
+import {
+  PlaceholderRaggedText,
+  PlaceholderText,
+  ProvidePlaceholderContext,
+} from "app/utils/placeholders"
+import { Separator, Spacer } from "palette"
+import { ActivityIndicator } from "react-native"
+
+export const BelowTheFoldPlaceholder: React.FC<{}> = ({}) => {
+  return (
+    <ProvidePlaceholderContext>
+      <Separator />
+      <Spacer mb={3} />
+      {/* About the artwork title */}
+      <PlaceholderText width={60} />
+      <Spacer mb={2} />
+      {/* About the artwork copy */}
+      <PlaceholderRaggedText numLines={4} />
+      <Spacer mb={3} />
+      <Separator />
+      <Spacer mb={3} />
+      <ActivityIndicator />
+      <Spacer mb={3} />
+    </ProvidePlaceholderContext>
+  )
+}
